refactor: 초반 IntroPage에서 navigate하는 로직을 App.tsx -> IntroPage로 이동
